move tests out of source, for now
