+++ __tests__/SaveToContact.test.tsx
@@ -2,8 +2,8 @@ import * as React from 'react';
 import { mount } from 'enzyme';
 
 import { SaveToContactComp } from '../components/actions/SaveToContact';
-import { FlowMutator } from '../components/FlowMutator';
-import { FlowDefinition } from '../FlowDefinition';
+import { FlowMutator } from '../src/components/FlowMutator';
+import { FlowDefinition } from '../src/FlowDefinition';
 import { getFavorites } from './utils';
 
 describe.skip('SaveToContact', () => {
