+++ __tests__/SwitchRouter.test.tsx
@@ -1,10 +1,10 @@
 import * as update from 'immutability-helper';
 import * as UUID from 'uuid';
-import { FlowMutator } from '../components/FlowMutator';
-import { FlowDefinition, Case, SwitchRouter, Exit, Node } from '../FlowDefinition';
+import { FlowMutator } from '../src/components/FlowMutator';
+import { FlowDefinition, Case, SwitchRouter, Exit, Node } from '../src/FlowDefinition';
 import { getFavorites, dump } from './utils';
-import { Config } from "../services/Config";
-import { CaseProps, resolveExits, CombinedExits } from "../components/routers/SwitchRouter";
+import { Config } from "../src/services/Config";
+import { CaseProps, resolveExits, CombinedExits } from "../src/components/routers/SwitchRouter";
 
 xdescribe('SwitchRouter', () => {
     let definition: FlowDefinition;
