+++ package.json
@@ -16,6 +16,7 @@
     "lint": "tslint -c tslint.json 'src/**/*.{ts,tsx}'",
     "build:dev": "webpack --env.NODE_ENV=dev --progress --profile --colors",
     "build:prod": "webpack --env.NODE_ENV=prod --progress --profile --colors",
+    "build:both": "npm run build:dev && npm run build:prod",
     "preversion": "npm test",
     "version": "npm run build:prod",
     "postversion": "git push && git push --tags && npm publish"
